+++ src/app/components/movie-item/movie-item.component.html
@@ -7,7 +7,9 @@
 		<button class="movie-card__button-watch" (click)="addToWatch()">
 			Add to Watch List
 		</button>
-		<div class="movie-card__rating">{{ transformedRating }}</div>
+		<div class="movie-card__rating">
+			{{ movie.vote_average | transformRating }}
+		</div>
 		<img
 			[src]="movie.poster_path"
 			[alt]="movie.title"
