+++ src/app/components/movie-item/movie-item.component.ts
@@ -1,39 +1,17 @@
-import {
-	Component,
-	EventEmitter,
-	Input,
-	Output,
-	PipeTransform,
-} from '@angular/core';
-
-// оголосив Pipe тут тому що не зрозумів як це зробити не додаючи модуль (я не хотів додавати модуль)
-export class TransformRatingPipe implements PipeTransform {
-	transform(value: number): string {
-		if (value % 1 === 0) {
-			return value.toFixed(0);
-		} else {
-			return value.toFixed(1);
-		}
-	}
-}
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { TransformRatingPipe } from '@pipes/transform-rating.pipe';
 
 @Component({
 	selector: 'app-movie-item',
 	standalone: true,
 	templateUrl: './movie-item.component.html',
 	styleUrls: ['./movie-item.component.scss'],
-	providers: [TransformRatingPipe],
+	imports: [TransformRatingPipe],
 })
 export class MovieItemComponent {
 	@Input() movie: any;
 	@Output() addFavorite = new EventEmitter<any>();
 	@Output() addWatch = new EventEmitter<any>();
-
-	constructor(private transformRatingPipe: TransformRatingPipe) {}
-
-	get transformedRating(): string {
-		return this.transformRatingPipe.transform(this.movie.vote_average);
-	}
 
 	addToFavorites() {
 		this.addFavorite.emit(this.movie);
