+++ src/app/pipes/transform-rating.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { TransformRatingPipe } from './transform-rating.pipe';
+
+describe('TransformRatingPipe', () => {
+  it('create an instance', () => {
+    const pipe = new TransformRatingPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
